RPC module
RPC module includes:
-ServiceAPI framework
-Json client/server
-TCP endpoint (client/server)

To allow a method for RPC, simply put add RpcAPIMethod meta attribute.
To name the API in which this method is, add RpcAPIService("ApiName") -> RemoteCall("ApiName.MethodName")
If RpcAPIService class attribut is omitted, the RPC methods will be in an anonymous API -> RemoteCall("MethodName")

This is the first draft. It support only intrinsic types, the server is not bullet-proof and the client is not "keep-alive' compatible (connection is reseted at each calls)

Next step is to support arrays and objects, keep-alive server and AnonymousPipes.